Incomplete View Commit

+++ src/main/java/driver/bookstore/Repository.java
@@ -4,7 +4,7 @@ package driver.bookstore;
 import javax.persistence.Entity;
 
 public interface Repository {
-    Entity findEntity(Long id);
+    Entity findEntity(int id);
     Entity updateEntity(Entity entity);
     void deleteEntity(Entity id);
     void addEntity(Entity entity);
